+++ react/components/atoms/button-tab-text/button-tab-text.js
@@ -1,5 +1,5 @@
 import React, { PureComponent } from "react";
-import { Platform, StyleSheet, Text, ViewPropTypes } from "react-native";
+import { Platform, StyleSheet, Text, View, ViewPropTypes } from "react-native";
 import PropTypes from "prop-types";
 import { capitalize } from "ripe-commons-native";
 import { mix } from "yonius";
@@ -12,9 +12,10 @@ export class ButtonTabText extends mix(PureComponent).with(IdentifiableMixin) {
     static get propTypes() {
         return {
             active: PropTypes.bool,
-            backgroundColor: PropTypes.string,
             color: PropTypes.string,
             colorSelected: PropTypes.string,
+            backgroundColor: PropTypes.string,
+            backgroundColorSelected: PropTypes.string,
             text: PropTypes.string,
             disabled: PropTypes.bool,
             activeOpacity: PropTypes.number,
@@ -28,9 +29,10 @@ export class ButtonTabText extends mix(PureComponent).with(IdentifiableMixin) {
     static get defaultProps() {
         return {
             active: false,
-            backgroundColor: "#f6f7f9",
             color: "#c8cdd2",
             colorSelected: "#00435e",
+            backgroundColor: "#f6f7f9",
+            backgroundColorSelected: "#4f7af8",
             text: undefined,
             disabled: false,
             activeOpacity: 0.5,
@@ -41,16 +43,39 @@ export class ButtonTabText extends mix(PureComponent).with(IdentifiableMixin) {
         };
     }
 
-    _style() {
+    _style = () => {
         return [
             styles.buttonTabText,
             styles[`buttonTabText${capitalize(this.props.variant)}`],
             { backgroundColor: this.props.backgroundColor },
+            this.props.variant === "colored"
+                ? {
+                      backgroundColor: this.props.active
+                          ? this.props.backgroundColorSelected
+                          : this.props.backgroundColor,
+                      borderTopColor: this.props.active
+                          ? this.props.backgroundColorSelected
+                          : "#e4e8f0",
+                      borderBottomColor: this.props.active
+                          ? this.props.backgroundColorSelected
+                          : "#e4e8f0",
+                      borderLeftColor: this.props.active
+                          ? this.props.backgroundColorSelected
+                          : "#e4e8f0",
+                      borderRightColor: this.props.active
+                          ? this.props.backgroundColorSelected
+                          : "#e4e8f0"
+                  }
+                : {},
             this.props.style
         ];
-    }
+    };
 
-    _styleText() {
+    _touchableStyle = () => {
+        return [styles.touchable];
+    };
+
+    _textStyle = () => {
         return [
             styles.text,
             styles[`text${capitalize(this.props.variant)}`],
@@ -58,39 +83,45 @@ export class ButtonTabText extends mix(PureComponent).with(IdentifiableMixin) {
                 color: this.props.color
             },
             this.props.disabled ? styles.textDisabled : {},
-            this.props.active ? { ...styles.textSelected, color: this.props.colorSelected } : {}
+            { color: this.props.active ? this.props.colorSelected : this.props.color }
         ];
-    }
+    };
 
     render() {
         return (
-            <Touchable
-                style={this._style()}
-                disabled={this.props.disabled}
-                activeOpacity={this.props.activeOpacity}
-                onPress={this.props.onPress}
-            >
-                <Text style={this._styleText()} {...this.id("button-tab-text")}>
-                    {this.props.text}
-                </Text>
-            </Touchable>
+            <View style={this._style()}>
+                <Touchable
+                    style={this._touchableStyle()}
+                    disabled={this.props.disabled}
+                    activeOpacity={this.props.activeOpacity}
+                    onPress={this.props.onPress}
+                >
+                    <Text style={this._textStyle()} {...this.id("button-tab-text")}>
+                        {this.props.text}
+                    </Text>
+                </Touchable>
+            </View>
         );
     }
 }
 
 const styles = StyleSheet.create({
     buttonTabText: {
-        alignItems: "center",
-        flex: 1,
-        justifyContent: "center"
+        flex: 1
     },
     buttonTabTextCompact: {
-        flex: 0,
-        paddingVertical: 9,
+        flex: 1,
         paddingHorizontal: 16
     },
+    touchable: {
+        flex: 1,
+        alignItems: "center",
+        justifyContent: "center"
+    },
     text: {
-        marginTop: Platform.OS === "ios" ? 4 : 0,
+        marginTop: Platform.OS === "ios" ? 10 : 4,
+        marginBottom: 8,
+        marginHorizontal: 10,
         fontFamily: baseStyles.FONT_BOOK,
         fontSize: 16,
         letterSpacing: 0.25
@@ -101,6 +132,5 @@ const styles = StyleSheet.create({
     },
     textDisabled: {
         opacity: 0.4
-    },
-    textSelected: {}
+    }
 });