+++ react/components/molecules/tabs-text/tabs-text.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { storiesOf } from "@storybook/react-native";
-import { withKnobs, select } from "@storybook/addon-knobs";
+import { withKnobs, select, text } from "@storybook/addon-knobs";
 
 import { TabsText } from "./tabs-text";
 
@@ -14,16 +14,32 @@ storiesOf("Molecules", module)
             { text: "New", disabled: false },
             { text: "Past", disabled: true }
         ];
+        const tabsBackgroundColor = text("Background Color", undefined);
+        const tabsBackgroundColorSelected = text("Selected Background Color", undefined);
+        const tabsColor = text("Tab Text Color", undefined);
+        const tabsColorSelected = text("Selected Tab Text Color", undefined);
         const variant = select(
             "Variant",
             {
                 Unset: undefined,
-                Compact: "compact"
+                Compact: "compact",
+                Colored: "colored"
             },
             undefined
         );
 
         const onTabChange = tabIndex => alert(`Switched to index: ${tabIndex}`);
 
-        return <TabsText tabs={tabs} tabSelected={0} variant={variant} onTabChange={onTabChange} />;
+        return (
+            <TabsText
+                tabs={tabs}
+                tabSelected={0}
+                tabsBackgroundColor={tabsBackgroundColor}
+                tabsBackgroundColorSelected={tabsBackgroundColorSelected}
+                tabsColor={tabsColor}
+                tabsColorSelected={tabsColorSelected}
+                variant={variant}
+                onTabChange={onTabChange}
+            />
+        );
     });